configMetaGenerator and  lastConfigMetaGenerator should be int64

+++ pkg/apis/kubedirector/v1beta1/kubedirectorcluster_types.go
@@ -30,7 +30,7 @@ type KubeDirectorClusterSpec struct {
 	ServiceType         *string     `json:"serviceType"`
 	Roles               []Role      `json:"roles"`
 	DefaultSecret       *KDSecret   `json:"defaultSecret"`
-	ConfigMetaGenerator int32       `json:"configMetaGenerator"`
+	ConfigMetaGenerator int64       `json:"configMetaGenerator"`
 	Connections         Connections `json:"connections"`
 }
 
@@ -52,7 +52,7 @@ type KubeDirectorClusterStatus struct {
 	ClusterService          string       `json:"clusterService"`
 	LastNodeID              int64        `json:"lastNodeID"`
 	Roles                   []RoleStatus `json:"roles"`
-	LastConfigMetaGenerator int32        `json:"lastConfigMetaGenerator"`
+	LastConfigMetaGenerator int64        `json:"lastConfigMetaGenerator"`
 }
 
 // +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
